Bones ...Again
Did exactly the same as last time but this time made a class for lines and points.   Only things left to do are convex Hull and Largest empty circle with their respective test files.  I did kind of half ass the intersection function for lines so if you can make a better algorithm to find if two line segments intersect go ahead.